requirements.txt
aiogram
pytz
requests
qrcode
Pillow
aiohttp
pyzbar
numpy
opencv-python-headless